requirements.txt
# GFlowFuzz_v2 - Complete Requirements
# Includes both base application and PPO/GRPO training dependencies with Unsloth optimization

# === BASE APPLICATION DEPENDENCIES ===
# MCP and API clients
mcp>=1.0.0
httpx>=0.24.0
openai>=1.0.0
requests>=2.31.0

# Configuration and utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
pydantic>=2.0.0
asyncio-extras

# Rich UI for terminal interface
rich>=13.0.0

# === PPO TRAINING DEPENDENCIES ===
# Core ML libraries
numpy>=1.24.0

# Hugging Face stack
transformers>=4.44.0
trl>=0.23.0
peft>=0.10.0
bitsandbytes>=0.43.1

# Progress tracking
tqdm>=4.65.0

# Data processing
datasets>=2.14.0

# Optional accelerate for distributed training
accelerate>=0.21.0

# Development and debugging
tensorboard>=2.13.0
